ZERO-92 [fix] : 리팩토링 및 비밀번호 검증 로직 추가
