+++ src/components/myhistory/MyTask.tsx
@@ -7,13 +7,8 @@ import Task from '@/src/types/myhistory/TaskType';
 import { useDeleteTask } from '@hooks/myhistory/useDeleteTask';
 import { useState } from 'react';
 import SideBar from '@components/@shared/SideBar';
-import Image from 'next/image';
-import networkErrorIcon from 'public/icons/networkErrorIcon.png';
-import Button from '@components/@shared/Button';
+import NetworkError from '@components/@shared/NetworkError';
 
-const handleReload = () => {
-  window.location.reload();
-};
 export default function MyTask() {
   const [isSideBarOpen, setIsSideBarOpen] = useState(false);
   const mutation = useDeleteTask();
@@ -25,20 +20,7 @@ export default function MyTask() {
   }
 
   if (isError) {
-    return (
-      <div className="flex flex-col items-center justify-center">
-        <Image
-          src={networkErrorIcon}
-          alt="네트워크 에러 아이콘"
-          width={100}
-          height={100}
-        />
-        <span className="text-2xl-bold">네트워크 에러</span>
-        <Button onClick={handleReload} className="mt-4 bg-amber-400">
-          재시도
-        </Button>
-      </div>
-    );
+    return <NetworkError />;
   }
 
   if (tasks.length === 0) {
